Build potential particle wall neighbors
In the particle wall handler relate bins to column wall elements being
stored a vector that is indexed via the lid of the element column map.

The potential owned particle to column wall element neighbor pairs are
build using aforementioned vector. A column wall element is seen as a
potential neighbor to an owned particle, if it is closer than the
minimum relevant bin size to that particle.

Access to the resulting potential neighbor pairs is granted via a method
defined in the particle wall handler interface.

Safety flags to check the validity of the relations are included.

part of #191