Remove deprecated functions of binary trees
For a list of removed functions please see issue #181.

The separate binary tree methods are renamed as the combined methods
are removed and therefore no distinction is needed any more. A
comment is added that there has also been a combined algorithm that got
removed as it is outperformed by the remaining one for large problems

Comments are changed accordingly.

Removed some unnecessary comments where not needed for beauty

part of #181