Remove couette flow SPH test case
The couette flow SPH test case is removed, because it is very sensitive
to round-off errors. In the past the results of this test case were
adapted very often. There is no functionality in this test case that is
not covered by the poiseuille flow or driven cavity test cases or unit
tests, thus it can be removed without any concerns.