Add thermal capacity to SPH thermo material
Add a material parameter for thermal (heat) capacity to the SPH thermo
material (related to #98). In addition the inverse of the thermal
capacity is stored so that a lot of divisions can be avoided in the
evaluation of the energy equation.

Including a test case covering the changes.

All material parameters needed for thermo evaluations are now optional,
adapting the test cases correspondingly.