Add heat source handler to SPH
Add a heat source handler to the SPH framework. The heat sources are
evaluated as a part of the SPH temperature evaluation scheme. Currently,
volume heat sources are implemented. Surface heat sources are
implemented later on if needed.

Extend thermal particle material by parameter of thermal absorptivity.

A test case is added that covers the new functionality.