Improve DirectDivergence RefPlane
In the DirectDivergence numerical integration approach for arbitrary
polyheadra a Reference Plane has to be choosen which influences the
resulting integration points in the volume.
For FEM we want the integration points to be inside the element, and
therefore a Reference Plane has to be choosen which guarantees that all
integration points are inside.

As this was not possible in the current version for general
configurations, the following modifications have be made:
        - Add a strategy to define reference plance based on volumecell
          facets
        - Do a more precice check if integration points will lie in the
          element (use corner points of volumecell instead of whole
element)
        - I the case we cannot find a plane which results guarantees
          that all integration points lie inside the element, we allow
for a deviation from that (in steps to get still the best choice)

Introduce the parameter "DIRECT_DIVERGENCE_REFPLANE in the "CUT GENERAL"
Section of the *.dat file to control the choice of the reference plane:
        - default is "all", to get a similar version to the old
          implementation choose "diagonal_side", additionally all
strategies can be set seperately

Modification to drt_fem_general:
        - Add bool-parameter "throw_error" to disable the dserror in
          case gausspoints are slightly outside the element (This
dserror was commented out by the CutFEM people already for all bigger
computations)

Modification to testcase:
        - As a different choice of Referenceplane influences the
          integration error, this can change the results. To keep the
original results for the testcase
xfluid_moving_torus_NavSlip_vareps_levelset_psmoothed_8x8x16, the
parameter DIRECT_DIVERGENCE_REFPLANE is set to diagonal_side.