Add temp grad driven surface tension to SPH
Adding a temperature gradient driven surface tension to the SPH
framework. The surface tension is assumed to show a linear temperature
dependency with negative slope. A positive scalar input parameter
defines the magnitude of the derivation of the surface tension w.r.t.
the temperature. The negative sign is accounted for in the
implementation.

A test case covering the new functionality is included.