Determine particle types from input parameter
So far the particle types to initially set up the particle containers
were determined from all initial particles (as defined in the input
file). After adding the SPH phase change handler (refer to #133) it
could happen that particles of a specific type were not given in the
input file as initial particles but generated during the simulation.
This issue was up to now circumvented by defining a particle of that
phase in the input file that is located outside of the bounding box and
thus removed during particle distribution (problems with the restart
were still present!).

With this commit an input parameter is added that defines all particle
types needed for the initial setup of the particle containers, thereby
resolving the problems described above.

All particle test cases are touched, however no changes in the results.
In the test cases related to SPH phase change (added with #133) the
global IDs of the particles have changed, as the initial dummy particle
(located outside of the bounding box) can now be removed.