Modify dynamic load balancing of particles
The bin weights needed for dynamic load balancing are determined
weighting the particles according to their particle type.

This enables to better distribute the work load on the processors, e.g.,
boundary particles in SPH are often computationally less expensive than
normal fluid particles.

The input parameter introduced in the previous commit is used to define
the weight of each particle type.

part of #115