Store recent communication of particles
The particle engine is responsible for communication of owned particles
from processor to processor, e.g., if particles travel over a processor
boundary or are redistributed during dynamic load balancing.

A data type is now added to the particle engine, that keeps track of the
recent communication of particles, storing which particle (global ids)
are send to which processor. Access is granted to that data type via the
particle engine interface.