Add particle wall result test handler
Add a particle wall result test handler that can check the position and
displacements of wall nodes, and the total number of wall elements and
nodes over all processors.

Add two test cases checking if the particle wall are set up correctly
for both types of particle wall source 'bounding box' and 'discret
condition' as introduced in the first part of #191 with !148.