Add method to transfer particle wall elements
Add a method to the particle wall handler that transfers wall elements
and nodes to their new processors based on the underlying distribution
of the bins. Thanks to @jeichinger for his work in the binning strategy
allowing me to re-use some of his methods.

In the class particle algorithm it is now also determined based on the
position increment of the wall nodes if a transfer of particles and
walls is needed.

Note: The wall displacement states (based on row and column map) are
added as members to the particle wall handler, but currently not set
yet. This is then done in the future via the PaSI framework (that still
needs to be migrated as part of milestone:
https://gitlab.lrz.de/baci/baci/milestones/3

part of #191