Update QHull include path
The qhull include path specified via QHULL_LIBRARY_DIR is now set in
each build configuration since it depends on the directory structure of
the TPL installation. These are different at LnM and IMCS.

These changes just alter configure variables and the only occurence of
a qhull header inclusion in the baci source code. TPLs don't need to be
recompiled.

Part of #224