Fix fsi_dc3D_part_ait_ga_ost_xwall test
The strcuture and fluid sovers were unconverged in the
old time integration framework. The solver parameters,
like the maximum number of iterations, were provided
differently within the xml file for the new time
integration.
To fix this issue, the solver was run until convergence
within the old time integration and those values are
now used for the test within the new time integration
related to #228