Change cmake macro Baci_framework_test
An additional argument names xmlfilename was introduced.
If a xml filename is given, it is copied from the
input firectory to the build directory in order to be used by the
dat file generated in the build directory.
I also changed the argument called arg to testname for it to be
easily understood and added some documentation.
related to #228