Test cases stuck in the old time integration
constr2D_fsi does not work in the new time integration
because the newtonlinuzawa solver is not implemented (or planned) to
be implemented there yet.
fs3i_ac_prestress is stuck because a monolithic atherosclerosis
problem is restartet from its results. And as of right now,
neither monolithic fsi problems not Windkessel models nor prestressing
is fully migrated to the new structural time integration.
related to #228