Reformulate the calculation of linearizations
Evaluation of Cauchy-stresses and their linearizations are rewritten
such that the linearizations are now formulated w.r.t. the deformation
gradient F, instead of the left Cauchy-Green tensor b.
This was required as the derivation for the linearizations that will
be introduced in a subsequent commit would require a singular value
decomposition (see issue #281).

Added override specifier and removed virtual where possible.

Made LINALG::Matrices static where possible.

Added a new function: AddRightNonSymmetricHolzapfelProductStrainLike
to be able to evaluate this product in strain-like Voigt notation.

part of #281