Remove unreasonable test quantities of test case
This test case is designed such that a compression in x-direction takes
place. Displacements in lateral directions at the surfaces are
restricted using DBCs. Therefore it is clear that displacements in
y- and z-directions are only very small and it is not reasonable to
test them and are therefore removed.

part of !204