Check for valid particle neighbors separately
In the case of no particle interactions the method
`HaveValidParticleConnectivity()` returns always false (resulting in a
particle transfer in every time step even if actually not needed) as the
list of potential particle neighbors is never build in that case. To
account for that the method is split up and the validity of the particle
neighbors is checked in a separate method only if particle interactions
are evaluated.