Remove testcase constr2D_fsi
According to the Baci steering committee, as of now, there is no plan to
migrate the Uzawa solver functionality to the new time integration.
Thus, in order to fully migrate the partitioned FSI framework to the new
time integration, it becomes necessary to remove the affected test case.

related to #228, #229