Fix bug in method GEO::nearest3DObjectOnElement()
The method `nearest3DObjectOnElement` correctly computes the closest
distance of a given point from a surface element. However, in some
special cases the wrong object type of the found point on the element is
returned. This is for example the case if the given point is
perpendicular on an element edge and located directly above a node.
Hence the closest point is the respective corner node of that edge resp.
element.

The algorithm in above mentioned method so far detected the correct
closest point (namely the node) but returned the object type for an edge
and not a node.

The problem can be easily fixed by replacing the distance comparisons
from `<` to `<=`.

This bug was detected setting up an academic example and probably does
not take effect for real examples/simulations.

This method is currently only used by particle related methods. However
in this file other method do some similar distance calculations. So to
all @baci/baci_developers: If you are using one of these functions and
rely on a correct object type check if the function you are using is
doing the right thing!