Fix memory leak implicit  CLN to double
- allow only explicit CLN to double conversion (avoid this memory leak
  in future)
- include clnwrapper.H !!!before!!! linalg_fixedsizematrix.H
  in affected headers

Background information:
Having ClnWrapper to double convert leads to caching of the
converted values in the
numbers cache if those values are later converted back to CLN (e.g.
in some arithemtical operations), that is supposed to be done only for
the compile time
constants. This eventually either exhausts custom memory manager or
leads
to memory fragmentation with normal allocator, since such values are
never freed. Therefore we now only allow explicit convert to
double values, for instance for easier printing and more control on
the behaviour.

Additionally to get proper look-up for ClnWrapper related overloads of
standart library math functions we are forced to include
cut_clnwrapper.H prior to inclusion of
src/linalg/linalg_fixedsizematrix.H so the template functions that are
using overloads can be properly instantiated.