Mortar: output of interface discretizations
As each mortar meshtying/contact interface holds its own
DRT::Discretization, it can use its own DRT::DiscretizationWriter to
write output to the disk, which then can also be post-filtered with the
ensight filter.

For now, this is optional and can be used as additional visualization
output. Restart data is still handled via the underlying structure
discretization.

The interfaces need to know about their displacements, so we have to
pass it (and some other data) from the time integration and
contact/meshtying managers to the interface. This is done via
Teuchos::ParameterLists.

Enabled interface-based output in 2D/3D meshtying and contact tests.
Unfortunately, the post filter can't be tested right now.

Limitations:
- Currently only implemented for the old time integration. It can be
hooked up with the new structure framework, but I haven't done that yet.
- Does not work for frictional contact, only for meshtying and
frictionless contact.

Part of #135.