Implement no physics ImplType for SSI #268
The implementation type does not solve any equations on the elements.
It could be used for boundary elements (BELE) that are needed only for
the structure, during a one-way ssi (scatra-to-solid). Currently, only
the minimal set of actions is implemented needed for a ssi simulation
with the scatra results from a restart file.