Nonlinear material models implemented
Nonlinear material models incorporating polynomials are implemented, which is mentioned in issue #255.
Test cases are updated because of the modifications in existing linear models (in consultation with @cschmidt).
By this modification, C_max is now chosen consistently in the code w.r.t. user input by 'MATID'.
Results of some test cases are changed due to rounding of factors (in consultation with @cschmidt).
Two new classes are introduced to avoid code duplication. They include functions which are commonly used
in various routines of child classes. So, the file becomes easier to maintain in future.