Check for neighboring particle-wall pairs in SPH
In the smoothed particle hydrodynamics (SPH) framework a method is added
that checks for particle to wall neighbor pairs, meaning a contact point
on the wall element is within the interaction distance of the particle.

This method is very similar to the corresponding method in the discrete
element method (DEM) framework being added with commit 39401f17 as part
of #238.

Note:
- In some cases the closest contact point of a particle to a wall
  element is not located inside the element but on an edge or a node. If
  several wall elements share an edge or a node this contact point
  should only be detected once!
- Contact points on an edge or a node inside the penetration volume of a
  surface contact point are not evaluated. The same accounts for contact
  points on a node inside the penetration volume of a line contact
  point.

part of #285