Fix initial stick flag in tangential contact (DEM)
The stick flag in the tangential contact evaluation needs to be
initialized to true whenever a new contact pair is detected and
evaluated.

This fix leads to slightly different results in a test case, that are
therefore adapted correspondingly.

Additionally fix a comment in the tangential contact handler.

Refer to #141, !176