Add rolling contact evaluation to DEM
Add a rolling contact evaluation handler to the DEM framework. Two
different types of rolling contact evaluation schemes, namely based on a
viscous and a coulomb law, can be selected via a corresponding input
parameter.

History pairs are handled very similar than for the tangential contact
evaluation, refer to #141 and !176.

Unit tests and test cases are included.

part of #286