Adaptive Evaluation of CUT in XFSI/XFPSI
Depending on the unscaled maximal displacement increment in the
in each Newton iteration, the evaluation of the CUT can be
avoided and the previous geometric intersection can be used.

This reduces the number of CUT evaluations and allows to reduces
problems with osscilating DOFsets (as they are purely geometry based)
during the Newton iterations.

- Modified testcase to test this new functionality: no change in result

- XFluid now stores additionally the state of the previous intersection

- Use input parameters: CUT_EVALUATE_MINTOL, and CUT_EVALUATE_MINITER
  to control the evaluation of the CUT --> Default is evaluate
  always.