Remove NLNSOL framework
The NLNSOL nonlinear solver framework was an experimental piece of code
to work towards nonlinear multi-level solvers in Baci. It never worked
properly and is not developed actively anymore. In order to reduce the
amount of code to maintain and to remove interfaces to MueLu, this code
is removed.

If a similar capability is required in the future, I think it should be
tackled by a different implementation approach.

Precedes #205.