Get particles within a search radius
Add a method to the particle engine that returns particles within a
search radius around a search point.

Note that in case the search point is located in a ghosted bin on this
processor, particles within the search radius but not owned or ghosted
by this processor are not found.

part of #285