Mortar: update master/slave sets with given maps
For nodes and elements, column maps can be given to be used to update
master/slave maps. Otherwise, the maps currently stored in the
discretization are used.

This change is necessary due to "extending ghosting via binning". In
this case, we need to update master/slave maps at a point, where the
discretization is not in a final FillComplete() state. We have new maps
and want to use those.

Part of #295, #307.