Reduce runtime of beam3sk lineload tests (#55)
Reduce number of time steps from 200 to 110.
Without an entire redesign of the tests, this is the most reasonable
choice since the applied dynamic line load is removed after 100 steps
and the beam swings freely after that.

The actual reason why these tests take so long is that they use
automatic differentiation (FAD) because these variants of beam elements
lack an analytic linearization. Rename the tests accordingly to reflect
this at first sight.