Add a check whether to skip evaluation
of beam-beam potential-based interaction pairs.

The idea is to get a conservative lower bound estimate for the minimal
centerline separation of both elements, which doesn't require any
interpolation, i.e., information on the deformation of the elements and
is thus cheap to compute before we begin with the actual, expensive
evaluation of the pair on Gauss point level.

All beam potential tests pass without change of results.
Runtime reduction is up to 30% for SBIP approach and 45% for SSIP
approach for the existing tests.