Contact: recreate ghosting only if redistributed
Ghosting via binning has only to be redone, if the interfaces have been
redistributed. Otherwise, the maps have not changed, so we don't need to
recompute the extended column maps.

This avoids a call to FillComplete() with the expensive operation of
assigning degrees of freedom.

Note: This is a preliminary fix. Once binning has been moved inside
ExtendInterfaceGhosting() for contact problems as well, this call to
bining will be removed anyways.

Part of #307.