Mortar: call FillComplete() only if necessary
A full and expensive FillComplete(), i.e. with assigndegreesoffreedom =
true, is only required on a discretization with a final parallel
distribution. We will not call FillComplete() with
assigndegreesoffreedom = true for mortar interface discretizations, that
are a priori known to be redistributed.

This enhancement has been implemented for solid meshtying and all mortar
adapters. Other problems, like wear or scatra, maintain backwards
compatibility by just passing 'true' as the respective function
argument. In doubt, previous behaviour is recovered, where
FillComplete() is called once too many.

Related to #323.
Part of #295, #307.