Introduce input parameter initial particle spacing
Introduce an input parameter for the initial particle spacing in SPH
that is used to initialize the initial particle volume. So far, a
problem volume was given via input parameter and divided by the total
number of non-boundary or non-rigid particles in the system to get the
initial particle volume. This approach is rather unhandy and leads to
errors, as can be seen from commit 3520f381.