Implemented scatra element boundary parameter
class. This has been done within this commit since so far the
corresponding values of the scatra-scatra interface condition had to be
read from the condition at each gauss point.

This is now changed as follows:
-Condition values are stored to implemented boundary parameter class at
 time integrator level before evaluate is called
-pointer to implemented boundary parameter class is stored to
 evaluation classes to have access to them
-required values are accessed using return methods

Results of three test cases with a lot of time steps have slightly
changed and were adapted. Change in results is probably caused by a
different round-off behavior as divisions are carried out a bit
different now. (I'm maintainer of those test cases)

part of #344