SelfCUT Update
1. The SelfCUT propagates its position only through Cutsides (as it was
inteded to be)
2. Fix handling of removed sides in Selfcut, now we do not get a
SegFault anymore when you access a side which was removed as it is
stored in a storagecontainer. Add these functions to the
CUT::Meshhandle.
3. Use option selfcut_do_meshcorrection_ to specify if the removal of
Island in the SelfCUT should be performed or not.
4. Add some quick checks in the Collision Search of the SelfCUT
5. Do not use very small sides in the SelfCUT to detect inside or
outside position (use the next one instead)
6. Completely rework the FindIslands() routine (which was not working
before...)
    --> We collect all physically connected sides on the interface and
decided depending on the size of the bounding box if we remove this from
the interface or not.

======================================================================

New Input parameters in the ---CUT GENERAL SECTION
- DO_SELFCUT (bool): active/deactive SelfCut
- SELFCUT_DO_MESHCORRECTION (bool): activate/deactivate meshcorrection
  in SelfCUT (FindIslands routine)
- SELFCUT_MESHCORRECTION_MULTIPLICATOR (int): multiple of interface
  element size as maximum size to remove Island in the meshcorrection