Introduce XFSCI by Nitsche
This enables XFSCI based CutFEM and the Nitsche's method.
What has to be done from an algorithmic point is the evaluation of
Nitsche-FSI and Nitsche-Contact type integrals on the entire interface
(boundary of the solid). As having one framework for the evaluation of
both types of terms is not so easy in BACI, exemplary issues:
- one need geometric operations such as the projection to the other
  contact partner and the other need the connection to the underlying
fluid ...
- the element matrixes have different lineariztion degrees of freedom
  (of the contact partner and one of the underlying fluid ...
this is done in the classical Nitsche XFEM interface integration
framework and classical Nitsche Contact framework of BACI.

As a result, XFEM and Contact need to talk to each other on a Gauss
point level. This is the job of the new object the
Xfluid-Contact-Communicator: XFluid_Contact_Comm, who lives in the
FSI_XFEM_Monolithic class in the case of XFSCI. XFEM can ask it for the
contact state and vice versa ...

To ensure continuity of the algorithm with respect to the numerical
integration on the interface all interface elements which are mixed
(have contact and FSI gausspoints) the integration points from the CUT
are used.

To use XFSCI the contact interface has to be specified by the condition
type "navslip_contact".

All other changes are related to the fact that XFSI now can handle more
challenging nonlinear problems which requir(ed)/(s) some adaptions on
various places of the code.

There are single places (especially in the xfluid contact communicator)
where small aspects are already there in preparation for XFPSCI fully
Nitsche based ... which will come soon.

-Introducde testcase to test the new functionality: xfpsci_towelement

Contact related modifications:
1. Communicate SearchElement to all Processors in the SelfContact
Binarytree (as we potentially have to evaluate contact in XFSI based on
the processors of the fluid domain)
2. Introduce contact_nitsche_strategy_fsi and
contact_nitsche_integrator_fsi which handle this case from the contact
side

CUT related modifications:
2. Update position in the Boundarycell as for the FrozenCut in
combination with Contact the FSI interface and the Contact interface
still have to match geometrically (for pure XFSI this is not critical)
3. Added new CutTest