Adapt documentation: drt_particle_interaction
Adapt the documentation in directory `drt_particle_interaction` to be
consistent to the style of documentation used in the rest of the
particle framework.

Note: All touched files are level 3 as still changes in code structure
may occur in the near future. Thus full doxygen documentation is not
added.

part of #91