Introduce XFPSCI Nitsche formulation
This commit introduces the CutFEM XFPI formulation including
Nitsche Contact.

This is extension of the Nitsche based XFSCI formulation.
Additionally the coupling terms for the porous flow
have to be treated.
Mainly two classes are introduced:

- The additional contact_nitsche_integrator_fpi integrates the
actual terms on the interface (or uses the implementation of
the corresponding base classes.)

- The additional contact_nitsche_strategy_fpi provides the
interface for the xfluid_contact_communicator.

Additional adaptions to be able to handle FSI and FPI in various
places.