Result checks 'invana_..._target.dat' tests (#354)
These test cases define the target discretization for the related
statistival inverse analysis test cases with surface current objective
function.
The tests are simple structure simulations.
We now check for the displacements at the tip of the arch.

Note: In case, we use the dat-files as target discretization for an
inverse analysis, we use the undeformed geometry to compare against.