Introduce flag to decide when scatra-scatra
interface meshtying is performed. Until now this has been done based on
the presence of a scatra-scatra interface coupling condition that
should only define the interface equation to be solved.
Of course this is not possible any more for the contact problem because
then the condition defining the interface parameter would imply that
there is mesh tying at this interface. This behavior is changed within
this commmit.

precedes #280