Include guards in fluid Neumann boundary calc
Bugfix: In the calculation of the fluid Neumann
boundary conditions, several values are
calculated but never used unless the "new" one
step theta time integration is used. This
calculations can lead to errors thrown in the
fucntin class, because a megative time is handed
in for some parameter sets of e.g. the
generelized alpha method. This sommit simple
guards these fucntion calls by an if clause. #215