Fix bug in particle engine
Fix a bug in the particle engine concerning particles leaving the
computational domain. In case the particle used to be in a bin that is
at the border of both the computational domain and also a neighboring
processor, an error is thrown although the particle is already marked to
be removed.