Get Baci running on Schmarrn
This commits renames all variables and references from
KAISER to SCHMARRN. Previously investigations have
shown that only baci-release is buildable on schmarrn.
When building the postprocessor, some obscure error is
shown that we couldn't resolve in a reasonable amount of time.

The variable to run a build test on schmarrn is
GITLAB_SCHEDULER_LNM_SCHMARRN.

The build command is make -j3