Allow optional particle radius as input parameter
Modify the particle input reader to allow for an optional particle
radius being read in.

This is currently only of interest in the case of
DEM simulations in combination with input flag `INITIAL_RADIUS` set to
`RadiusFromParticleInput`. Note the default option for `INITIAL_RADIUS`
is `RadiusFromParticleMaterial`.

In the case of SPH simulations the radius is always equal for all
particles and set from the particle material.