Hand stiffness contributions to the fluid
In the case of weak dirichlet couplign of the
fluid velocity, contributions to the stiffness
matrix have to be assebled in every newtin step.
In order to keeo the fluid and the solid separate
from each other, this is done y simply handing a
linear matrix to the fluid to be added onto the
syste matrix within every newtin step.
cf. #194