Add open boundary handler to SPH
An open boundary handler is added to the Smoothed Particle Hydrodynamics
(SPH) framework capable of prescribing Dirichlet and Neumann boundary
conditions.

Dirichlet boundary condition:
The idea is to prescribe the velocity field of the dirichlet particles
and to extrapolate the pressure from the fluid particles to the
dirichlet particles. The density results form evaluation of the equation
of state.

Neumann boundary condition:
The pressure is prescribed for neumann particles, the density results
form evaluation of the equation of state. The velocity field is not
prescribed so far, but the acceleration of the neumann particles is
evolved via the momentum equation.

Changes are covered by two new test cases!

part of #388