Introduce two scatra-scatra interface conditions
The first one introduces the possibility to apply a constant interface
resistance, accounting for a non optimal electrical contact at this
interface. The second one enables to specifiy a no-flux condition in
combination with an 'SSI INTERFACE MESHTYING' condition.

Both conditions are needed to model the current collector - active
material interface in all-solid-state batteries more realistic.

A test case is introduced to check this new functionality. All other
tests are unaffected.

part of #395