More robust Genalpha time integration parameters
The default values of the generalized-alpha time integration have been
changed.
ALPHA_M, ALPHA_F, BETA and GAMMA now have a default value of -1.0.
RHO_INF is now by default 1.0.
If the user wants to use RHO_INF, he must only provide a RHO_INF
within [0,1] in the INPUT file, then a optimal set of the four
parameters will be calculated and used for the time integration.
If the user wants to use the other four parameters directly, he must
provide ALL of them meanwhile expicitly set RHO_INF as -1.0.
This commit also includes a change which prevents the program to extract
ALPHA_F directly from the input, which might cause a bug when RHO is
used in the input.

part of #123