Add linalg unit tests: determinant
Add unit tests concerning the testing of the evaluation of determinants
in linalg utils.

Thereby, a bug in method `double LINALG::DeterminantSVD` was detected.
The determinant seems to have a wrong sign, refer to method `double
LINALG::DeterminantLU` where the sign is modified due to swapping of
rows.

The method is currently not called in BACI. Hence, a dserror is added
and the unit tests for this method have a leading "X" in their name, and
are thus not recognized by test discovery of CxxTest. As soon as the bug
is fixed, enable the unit tests!

part of #416