Save genhtml output to artifact print in pages job
Save the output of the genhtml command as an artifact of the
buildtest_coverage_weekly job and enable the printing of the current
coverage rate (line coverage) during the pages stage.
In a follow up, we will add the pages stage to all pipelines run ob
master.
This will fix the following issues:
- failing deployment due to out of date hash (happen if a commit to
master happens after the deploy pipeline has been started)
- unknown coverage rate in the coverage badge (happens if newer
pipelines finish after the weekly coverage pipeline that do not have a
buildtest_weekly_coverage job)