Fix deployment to GitLab Pages and Coverage Badge
Change the execution rules of the pages job in the pages stage of
pipelines such that they are triggered for all pipelines on the master
branch. This includes pipelines that are automatically triggered by
commits to the master branch as well as the scheduled pipelines (daily
release and debug pipelines).

With this we can ensure that the current weekly coverage report will be
uploaded to GitLab Pages. Additionally, the coverage badge will now show
the current weekly line coverage value, since the current value will be
printed in all pipelines on master during the pages job. Therefore it
can be detected by the coverage regex set in GitLab settings.

GitLab runner cache is used to pass the required data
(the coverage report and the line coverage value) between pipelines.
Namely, from pipeline that builds the coverage report to all master
pipelines (more specific: the pages job within the pages stage of all
master pipelines).