Enable initial output for mortar interface output
Mortar/contact interfaces can write their own output, but this requires
access to D and M matrices. If they are not available yet (as when
writing output at the initial state of a simulation), don't use them to
compute nodal forces (which are zero anyways). Just write a zero vector.

Follows !428