Increase robustness of coverage pipeline
Move the creation of the coverage report to the after_script section of
the weekly coverage job.
This ensures the execution also in case of failing test cases.
Failing test cases should not influence the result of the coverage
report to much (at least in the case of integration tests). For
measurement of the coverage we are more interested in which code is
executed and less in that it is correctly executed (since we test the
correct execution in daily pipeline anyways!!).

Additionally:
- introduce the measurement of the baseline coverage (see
documentation of lcov). This increases the accuracy of the coverage
measurement
- exclude external source files (like e.g. boost library) from the
coverage measurement.
- introduce commit hash in test title such that one knows on which
commit the coverage report is based
- clicking on the sha will link to the respective GitLab page of the
commit
- add .info files to the artifacts to be safe