Contact: switch test back to classical crosspoints
Slave nodes on boundaries can be converted to master nodes with an
appropriate modification of the Lagrange multiplier shape functions. We
have two implementations to do so. Switch back to the original one by
Alexander Popp, since the new one runs into problems with redistribution
and does not cover all cases anyways.

Part of #398.