Move some scatra element boundary calculation elch
electrode methods to the newly created utils file. This file will be
used in future to add all methods that might also be called from
outside the class it is currently defined in.

This issue came up during work on #280, where I will reuse one of the
methods I just moved to this new location.

Also made use of some c++11 features