Disable MueLu contact precs by preprocessor guard
MueLu contact preconditioners require more effort to be upgraded to a
new Trilinos version. It is not even clear right now, which parts will
be upgraded and which parts will just be replaced by existing MueLu
code. Hence, disable them via the define guard HAVE_MueLuContact, which
is not defined anywhere in any configuration for now.

Note: Code builds, but tests have not been run.

Part of #205.