IMCS: build configuration for Trilinos 12.14.1
Linking Baci against the new Trilinos version 12.14.1, internally
referred to as Q1_2019 (version from first quarter in year 2019) is
defined in a new build configuration file.

Due to changes in Trilinos' FindTPLHDF5.cmake, that now requires
parallel HDF5  if MPI is enabled for Trilinos, we need to manually
specify the paths to the parallel HDF5 libraries for some unknown
reason.

Part of #205.