Migrate explicit G&R for cylinder to mixture
This commit migrates the homogenized constrained mixture model for the
isotropic growth case for an cylinder to the mixture framework. The
results have been verified with the old framework. The new framework
better represents in the code what's going on physically. It is also
flexible. The current implementation is also twice as fast in the
explicit case as the tensors only needed for the implicit evaluations
are not computed.